Move public top-level extension functions and properties from core module to internal module
