+++ dd-sdk-android-core/api/dd-sdk-android-core.api
@@ -551,11 +551,6 @@ public final class com/datadog/android/core/SdkReference {
 	public final fun get ()Lcom/datadog/android/api/SdkCore;
 }
 
-public final class com/datadog/android/core/StrictModeExtKt {
-	public static final fun allowThreadDiskReads (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
-	public static final fun allowThreadDiskWrites (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
-}
-
 public final class com/datadog/android/core/UploadWorker : androidx/work/Worker {
 	public static final field Companion Lcom/datadog/android/core/UploadWorker$Companion;
 	public fun <init> (Landroid/content/Context;Landroidx/work/WorkerParameters;)V
@@ -790,21 +785,6 @@ public final class com/datadog/android/core/internal/utils/JsonSerializer {
 	public static final field INSTANCE Lcom/datadog/android/core/internal/utils/JsonSerializer;
 	public final fun safeMapValuesToJson (Ljava/util/Map;Lcom/datadog/android/api/InternalLogger;)Ljava/util/Map;
 	public final fun toJsonElement (Ljava/lang/Object;)Lcom/google/gson/JsonElement;
-}
-
-public final class com/datadog/android/core/internal/utils/MapUtilsKt {
-	public static final fun getNULL_MAP_VALUE ()Ljava/lang/Object;
-}
-
-public final class com/datadog/android/core/internal/utils/NumberExtKt {
-	public static final fun toHexString (I)Ljava/lang/String;
-	public static final fun toHexString (J)Ljava/lang/String;
-	public static final fun toHexString (Ljava/math/BigInteger;)Ljava/lang/String;
-}
-
-public final class com/datadog/android/core/internal/utils/ThreadExtKt {
-	public static final fun asString (Ljava/lang/Thread$State;)Ljava/lang/String;
-	public static final fun loggableStackTrace ([Ljava/lang/StackTraceElement;)Ljava/lang/String;
 }
 
 public final class com/datadog/android/core/metrics/MethodCallSamplingRate : java/lang/Enum {
