+++ dd-sdk-android-core/src/main/kotlin/com/datadog/android/core/internal/utils/MapUtils.kt
@@ -6,14 +6,6 @@
 
 package com.datadog.android.core.internal.utils
 
-// TODO RUM-373 public as hack, no other solution for now. Any?.toJsonElement relies on this
-//  particular value. Maybe create something like (class NullMap) and check identity instead?
-/**
- * Special value for missing attribute.
- */
-@Suppress("PLATFORM_CLASS_MAPPED_TO_KOTLIN", "PackageNameVisibility")
-val NULL_MAP_VALUE: Object = Object()
-
 /**
  * @return a new mutable map containing all key-value pairs from the given array of pairs.
  *
