+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/internal/anr/ANRDetectorRunnable.kt
@@ -10,8 +10,7 @@ import android.os.Handler
 import com.datadog.android.api.InternalLogger
 import com.datadog.android.api.feature.FeatureSdkCore
 import com.datadog.android.core.feature.event.ThreadDump
-import com.datadog.android.core.internal.utils.asString
-import com.datadog.android.core.internal.utils.loggableStackTrace
+import com.datadog.android.internal.utils.asString
 import com.datadog.android.internal.utils.loggableStackTrace
 import com.datadog.android.rum.GlobalRumMonitor
 import com.datadog.android.rum.RumAttributes
