+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/internal/instrumentation/gestures/GesturesUtils.kt
@@ -9,7 +9,7 @@ package com.datadog.android.rum.internal.instrumentation.gestures
 import android.content.Context
 import android.content.res.Resources
 import android.view.View
-import com.datadog.android.core.internal.utils.toHexString
+import com.datadog.android.internal.utils.toHexString
 import com.datadog.android.rum.tracking.InteractionPredicate
 import com.datadog.android.rum.tracking.ViewTarget
