+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/prerequisite/MemoryRequirementChecker.kt
@@ -7,10 +7,10 @@
 package com.datadog.android.sessionreplay.internal.prerequisite
 
 import com.datadog.android.api.InternalLogger
-import com.datadog.android.core.allowThreadDiskReads
 import com.datadog.android.core.internal.persistence.file.canReadSafe
 import com.datadog.android.core.internal.persistence.file.existsSafe
 import com.datadog.android.core.internal.persistence.file.readLinesSafe
+import com.datadog.android.internal.utils.allowThreadDiskReads
 import java.io.File
 
 internal class MemoryRequirementChecker(
