+++ reliability/single-fit/okhttp/build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
             )
         }
     }
+    testImplementation(project(":dd-sdk-android-internal"))
     testImplementation(testFixtures(project(":dd-sdk-android-core")))
     testImplementation(project(":reliability:stub-core"))
     testImplementation(libs.bundles.jUnit5)
