+++ reliability/single-fit/trace/build.gradle.kts
@@ -50,6 +50,7 @@ dependencies {
             )
         }
     }
+    testImplementation(project(":dd-sdk-android-internal"))
     testImplementation(testFixtures(project(":dd-sdk-android-core")))
     testImplementation(project(":reliability:stub-core"))
     testImplementation(libs.bundles.jUnit5)
